ai-service/requirements.txt
fastapi==0.114.0
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.32.3
black==24.8.0